git-svn-id: svn://123.57.173.36/meijia-parent@847 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/dao/xcloud/XcompanyDeptMapper.java
@@ -7,7 +7,7 @@ import com.simi.po.model.xcloud.XcompanyDept;
 public interface XcompanyDeptMapper {
     int deleteByPrimaryKey(Long deptId);
 
-    int insert(XcompanyDept record);
+    Long insert(XcompanyDept record);
 
     int insertSelective(XcompanyDept record);
